fix MMC1 mapper bug where the 16-bit unsigned integer size for mapped address inside cartridge mappers was too small, changed to size_t

+++ includes/mapper.h
@@ -8,11 +8,11 @@
 
 typedef struct mapper_t
 {
-   cartridge_access_mode_t (*cpu_read)  (nes_header_t *header, uint16_t position, uint16_t *mapped_addr, void* internal_registers);
-   cartridge_access_mode_t (*ppu_read)  (nes_header_t *header, uint16_t position, uint16_t *mapped_addr, void* internal_registers);
+   cartridge_access_mode_t (*cpu_read)  (nes_header_t *header, uint16_t position, size_t *mapped_addr, void* internal_registers);
+   cartridge_access_mode_t (*ppu_read)  (nes_header_t *header, uint16_t position, size_t *mapped_addr, void* internal_registers);
    // cpu write function takes in data parameter which is the data to write, this can be intercepted by cartridge mapper to do stuff like bank switching for example
-   cartridge_access_mode_t (*cpu_write) (nes_header_t *header, uint16_t position, uint8_t data, uint16_t *mapped_addr, void* internal_registers);
-   cartridge_access_mode_t (*ppu_write) (nes_header_t *header, uint16_t position, uint16_t *mapped_addr, void* internal_registers);
+   cartridge_access_mode_t (*cpu_write) (nes_header_t *header, uint16_t position, uint8_t data, size_t *mapped_addr, void* internal_registers);
+   cartridge_access_mode_t (*ppu_write) (nes_header_t *header, uint16_t position, size_t *mapped_addr, void* internal_registers);
    void (*init) (nes_header_t* header, void* internal_registers); // function to initialize a mapper's register if necessary
 } mapper_t;
